chore(e2e): move geometry-sensitive checks to MCP-only suite\n\n- Remove flaky overlap/card-bounds checks from CI suite\n- Add e2e-mcp/ with MCP-only spec and config\n- Add npm scripts: e2e, e2e:mcp\n- Ignore Playwright artifacts and e2e DB
