+++ package.json
@@ -10,6 +10,8 @@
     "test": "jest",
     "test:watch": "jest --watch",
     "test:coverage": "jest --coverage",
+    "e2e": "playwright test -c e2e/playwright.config.js",
+    "e2e:mcp": "playwright test -c e2e/playwright.mcp.config.js",
     "lint": "eslint . --ext .js",
     "format": "prettier --write AGENTS.md CLAUDE.md CONTRIBUTING.md docs/PROJECT_SCOPE.md .github/pull_request_template.md .github/workflows/ci.yml eslint.config.js .prettierrc.json README.md package.json",
     "format:check": "prettier --check AGENTS.md CLAUDE.md CONTRIBUTING.md docs/PROJECT_SCOPE.md .github/pull_request_template.md .github/workflows/ci.yml eslint.config.js .prettierrc.json README.md package.json"
@@ -25,9 +27,10 @@
     "sqlite-vec": "^0.1.7-alpha.2"
   },
   "devDependencies": {
-    "@eslint/js": "^9.9.0",
     "@babel/core": "^7.28.3",
     "@babel/preset-env": "^7.28.3",
+    "@eslint/js": "^9.9.0",
+    "@playwright/test": "^1.55.0",
     "babel-jest": "^29.7.0",
     "eslint": "^9.9.0",
     "jest": "^29.7.0",
